#506: Better structure and componentize our login/register/reset mechanisms
