+++ pages/login.vue
@@ -1,7 +1,9 @@
 <template>
   <b-container class="py-5">
     <b-card class="shadow">
-      <Login />
+      <Login
+        show-register-link
+        show-reset-link />
     </b-card>
   </b-container>
 </template>
@@ -11,6 +13,12 @@ import Login from "~/components/Login";
 export default {
   components: {
     Login
+  },
+  mounted() {
+    // If we are already logged in then lets redirect to dashboard
+    if (this.$auth.loggedIn) {
+      this.$router.push("/dashboard");
+    }
   }
 };
 </script>
